+++ src/library/filter/BlockFilter.h
@@ -5,15 +5,15 @@
 namespace PR {
 class PR_LIB BlockFilter : public IFilter {
 public:
-	explicit BlockFilter(size_t radius = 1)
+	explicit BlockFilter(int radius = 1)
 		: mRadius(radius)
 	{
 	}
 
-	size_t radius() const override { return mRadius; }
+	int radius() const override { return mRadius; }
 	float evalWeight(float, float) const override { return 1.0f / ((2 * mRadius + 1) * (2 * mRadius + 1)); }
 
 private:
-	size_t mRadius;
+	int mRadius;
 };
 } // namespace PR
