+++ src/library/filter/IFilter.h
@@ -7,10 +7,10 @@ class PR_LIB IFilter {
 public:
 	virtual ~IFilter() = default;
 
-	virtual size_t radius() const					 = 0;
+	virtual int radius() const					 = 0;
 	virtual float evalWeight(float x, float y) const = 0;
 
-	inline size_t width() const { return 2 * radius() + 1; }
-	inline size_t height() const { return width(); }
+	inline int width() const { return 2 * radius() + 1; }
+	inline int height() const { return width(); }
 };
 } // namespace PR
